Added appMap file for mapping programs to files
Also, custom.conf's syntax has changed to accomodate the programs that did not stick to the NameOfAppThenFileSpecificName file-naming convention that most of them use.

appMap's syntax is very easy to follow if you can deal with custom.conf's syntax. What you get will be dependent on the rest of the lines. Also, do NOT have line-wrapping enabled with any of these files, multi-line is not supported yet.